pyproject.toml
[tool.poetry]
name = "django-metronome"
version = "0.1.0"
description = "Django integration for Metronome billing"
authors = ["Gable Brown"]
readme = "README.md"
license = "MIT"
packages = [{ include = "django_metronome", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
django = ">=5.2,<6"
metronome-sdk = "^4.5.0"
pydantic = "^2.13.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0"
pytest = "^8.3"
pytest-django = "^4.11"
ruff = "^0.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
src = ["src", "tests", "example"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
# Generated by django-admin startproject; keep Django defaults.
"example/**" = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_site.settings"
pythonpath = ["example", "src"]
testpaths = ["tests"]
